# src/energy_optimizer/openapi-docs.yml
openapi: 3.1.0
info:
  title: Energy Optimizer
  version: 0.1.0
paths:
...
  /api/v1/dashboard/data:
    get:
      summary: Dashboard Data
      description: 'Return the versioned dashboard contract without mixing scenarios.


        The actual scenario returns the hourly historic series of every available

        asset, aligned to the requested half-open UTC range with null gaps, and

        lists in `assets` why any asset contributed no series. Absent, stale, or

        corrupt assets never invalidate the series of other assets.'
      operationId: dashboard_data_api_v1_dashboard_data_get
      parameters:
      - name: start_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Inclusive timezone-aware range start
          title: Start Time
        description: Inclusive timezone-aware range start
      - name: end_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Exclusive timezone-aware range end
          title: End Time
        description: Exclusive timezone-aware range end
      - name: scenario_kind
        in: query
        required: false
        schema:
          enum:
          - actual
          - forecast
          - plan
          - efficiency
          type: string
          default: actual
          title: Scenario Kind
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardDataResponse'
        '503':
          description: Dashboard data persistence is unavailable
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/dashboard/excluded-hours:
    get:
      summary: Excluded Hours
      description: 'List every hour excluded from imported history in a UTC range.


        An hour of Home Assistant history is imported only if every data point that

        contributes to it is valid. Every other hour has no value and is listed here

        with each cause: the entity, a reason code, a message, and the exact data

        points. Hours that Home Assistant no longer holds, because the service was

        down for longer than its history is retained, are listed with the reason

        `history_unavailable`; their cause names the missing range and has no entity

        and no data points. A source that is not configured, has no persisted history

        yet, or is corrupt is reported in `sources` and never hides the other sources.'
      operationId: excluded_hours_api_v1_dashboard_excluded_hours_get
      parameters:
      - name: start_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Inclusive timezone-aware range start
          title: Start Time
        description: Inclusive timezone-aware range start
      - name: end_time
        in: query
        required: true
        schema:
          type: string
          format: date-time
          description: Exclusive timezone-aware range end
          title: End Time
        description: Exclusive timezone-aware range end
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ExcludedHoursResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    DashboardAssetAvailability:
      properties:
        asset:
          type: string
          minLength: 1
          title: Asset
          description: Stable asset key
        status:
          type: string
          enum:
          - available
          - empty
          - stale
          - not_configured
          - unavailable
          - invalid
          title: Status
          description: 'available: series returned; empty: history exists but has
            no observation in the range; stale: series returned but the newest observation
            exceeds the polling threshold; not_configured: the installation has no
            source for this asset; unavailable: configured but no usable persisted
            data; invalid: persisted data is corrupt or could not be recovered and
            is withheld'
        series_ids:
          items:
            type: string
          type: array
          title: Series Ids
        reason:
          anyOf:
          - type: string
          - type: 'null'
          title: Reason
          description: Actionable explanation when status is not available
      additionalProperties: false
      type: object
      required:
      - asset
      - status
      title: DashboardAssetAvailability
      description: Whether one asset contributed data to a dashboard response, and
        why not.
    DashboardDataResponse:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
        status:
          type: string
          enum:
          - validated
          - partial
          - stale
          - empty
          - unavailable
          - invalid
          - infeasible
          title: Status
        requested_start_time:
          type: string
          format: date-time
          title: Requested Start Time
        requested_end_time:
          type: string
          format: date-time
          title: Requested End Time
        interval_minutes:
          type: integer
          const: 60
          title: Interval Minutes
        series:
          items:
            $ref: '#/components/schemas/DashboardSeries'
          type: array
          title: Series
        metrics:
          items:
            $ref: '#/components/schemas/DashboardMetric'
          type: array
          title: Metrics
        assets:
          items:
            $ref: '#/components/schemas/DashboardAssetAvailability'
          type: array
          title: Assets
          description: Per-asset availability; populated for the actual scenario
        diagnostics:
          items:
            type: string
          type: array
          title: Diagnostics
        plan_summary:
          anyOf:
          - $ref: '#/components/schemas/DashboardPlanSummary'
          - type: 'null'
      additionalProperties: false
      type: object
      required:
      - schema_version
      - status
      - requested_start_time
      - requested_end_time
      - interval_minutes
      - series
      title: DashboardDataResponse
      description: Versioned read contract for actual, forecast, and plan dashboard
        data.
    DashboardMetric:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        label:
          type: string
          minLength: 1
          title: Label
        value:
          type: number
          title: Value
        unit:
          type: string
          minLength: 1
          title: Unit
      additionalProperties: false
      type: object
      required:
      - id
      - label
      - value
      - unit
      title: DashboardMetric
      description: One scalar dashboard measurement displayed separately from diagnostics.
    DashboardPlanSummary:
      properties:
        status:
          type: string
          enum:
          - available
          - empty
          - infeasible
          - unavailable
          title: Status
        objective_value:
          anyOf:
          - type: number
          - type: 'null'
          title: Objective Value
        diagnostics:
          items:
            type: string
          type: array
          title: Diagnostics
      additionalProperties: false
      type: object
      required:
      - status
      title: DashboardPlanSummary
      description: Optional plan-level diagnostics shared by dashboard consumers.
    DashboardSeries:
      properties:
        id:
          type: string
          minLength: 1
          title: Id
        data_type:
          type: string
          minLength: 1
          title: Data Type
        scenario_kind:
          type: string
          enum:
          - actual
          - forecast
          - plan
          title: Scenario Kind
        timestamps:
          items:
            type: string
            format: date-time
          type: array
          title: Timestamps
        values:
          items:
            anyOf:
            - type: number
            - type: 'null'
          type: array
          title: Values
        unit:
          type: string
          minLength: 1
          title: Unit
        source:
          anyOf:
          - $ref: '#/components/schemas/SourceMetadata'
          - type: 'null'
        requested_start_time:
          type: string
          format: date-time
          title: Requested Start Time
        requested_end_time:
          type: string
          format: date-time
          title: Requested End Time
        available_start_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Available Start Time
        available_end_time:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Available End Time
        retrieved_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Retrieved At
        generated_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Generated At
        published_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Published At
        freshness:
          type: string
          enum:
          - fresh
          - stale
          - unknown
          title: Freshness
        validation_status:
          type: string
          enum:
          - valid
          - suspect
          - invalid
          title: Validation Status
        missing_intervals:
          items:
            type: string
            format: date-time
          type: array
          title: Missing Intervals
        is_default:
          type: boolean
          title: Is Default
          default: false
        calculation_status:
          anyOf:
          - type: string
            enum:
            - calculated
            - calculated_with_defaults
            - defaulted
            - unavailable
            - invalid
          - type: 'null'
          title: Calculation Status
      additionalProperties: false
      type: object
      required:
      - id
      - data_type
      - scenario_kind
      - timestamps
      - values
      - unit
      - requested_start_time
      - requested_end_time
      - available_start_time
      - available_end_time
      - retrieved_at
      - freshness
      - validation_status
      title: DashboardSeries
      description: One aligned, provider-independent dashboard series.
...
    ExcludedHoursResponse:
      properties:
        schema_version:
          type: string
          const: '1'
          title: Schema Version
        requested_start_time:
          type: string
          format: date-time
          title: Requested Start Time
          description: Inclusive range start
        requested_end_time:
          type: string
          format: date-time
          title: Requested End Time
          description: Exclusive range end
        excluded_hour_count:
          type: integer
          minimum: 0.0
          title: Excluded Hour Count
          description: Excluded hours summed over all sources
        sources:
          items:
            $ref: '#/components/schemas/ExcludedHoursSource'
          type: array
          title: Sources
        summary:
          items:
            $ref: '#/components/schemas/ExcludedHoursSummary'
          type: array
          title: Summary
        hours:
          items:
            $ref: '#/components/schemas/ExcludedHour'
          type: array
          title: Hours
          description: Excluded hours in ascending order of the hour
      additionalProperties: false
      type: object
      required:
      - schema_version
      - requested_start_time
      - requested_end_time
      - excluded_hour_count
      - sources
      - summary
      - hours
      title: ExcludedHoursResponse
      description: Every hour that was excluded from imported history in a requested
        range.
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError